requirements.txt
redis
flask
kafka-python
pyflink
gunicorn